Test fill added

+++ app/Http/Controllers/TestCategoryController.php
@@ -15,14 +15,15 @@ class TestCategoryController extends Controller
 
     public function create(Test $test)
     {
-        error_log("test in testcat create");
-        error_log($test);
+        //error_log("test in testcat create");
+        //error_log($test);
 
         if (Auth::user() == null || !Auth::user()->hasRole('profesor')) {
             return redirect()->route('home');
         }
         $categories = Category::all()->pluck('name', 'id');
 
+        error_log("jsem tu");
 
         return view('test_category.create', compact('test', 'categories'));
     }
@@ -41,9 +42,9 @@ class TestCategoryController extends Controller
             return redirect()->route('home');
         }
         $categories = Category::all();
-        error_log($category->id);
+        //error_log($category->id);
         $test_category = $test->categories->whereIn('id', $category->id)->first();
-        error_log($test_category);
+        //error_log($test_category);
 
 
         return view('test_category.edit', compact('test_category', 'categories', 'test'));
@@ -54,22 +55,30 @@ class TestCategoryController extends Controller
         if (Auth::user() == null || !Auth::user()->hasRole('profesor')) {
             return redirect()->route('home');
         }
+        $categoryQuestionNumber = Question::all()->whereIn('category_id', $request->category_id)->count();
+
+        if($categoryQuestionNumber < $request->number_of_questions) {
+            return redirect()->back()->withErrors(['error' => 'Number of questions is bigger than the actual number of questions in selected category']);
+
+        }
         $test->categories()->attach($request->category_id, ['number_of_questions' => $request->number_of_questions]);
 
-        Session::flash('message', 'Category questions created1 successfully');
+        Session::flash('message', 'Category questions created successfully');
 
         return redirect()->route('tests.edit', $test->id);
     }
 
     public function update(Request $request, Test $test, $category_id)
     {
-
-        error_log($request->number_of_questions);
-        error_log($request->number_of_questions);
-        error_log($request->number_of_questions);
-
         if (Auth::user() == null || !Auth::user()->hasRole('profesor')) {
             return redirect()->route('home');
+        }
+
+        $categoryQuestionNumber = Question::all()->whereIn('category_id', $category_id)->count();
+
+        if($categoryQuestionNumber < $request->number_of_questions) {
+            return redirect()->back()->withErrors(['error' => 'Number of questions is bigger than the actual number of questions in selected category']);
+
         }
 
         $test->categories()->detach($category_id);
